Moves ServiceInstanceSupplier to loadbalancer module
